refactor: move my-units pages under spa dir
