+++ apps/admin-ui/src/component/BufferToggles.tsx
