--- apps/admin-ui/src/component/my-units/create-reservation/utils.ts
@@ -1,28 +1,0 @@
-import { camelCase, get, pick, zipObject } from "lodash";
-import {
-  type ReservationFormType,
-  type RecurringReservationForm,
-  type ReservationChangeFormType,
-} from "@/schemas";
-import { type ReservationMetadataFieldNode } from "@gql/gql-types";
-
-export function flattenMetadata(
-  values:
-    | ReservationFormType
-    | RecurringReservationForm
-    | ReservationChangeFormType,
-  metadataSetFields: Pick<ReservationMetadataFieldNode, "fieldName">[]
-) {
-  const fieldNames = metadataSetFields.map((f) => f.fieldName).map(camelCase);
-  // TODO don't use pick
-  const metadataSetValues = pick(values, fieldNames);
-
-  const renamePkFields = ["ageGroup", "homeCity", "purpose"];
-
-  return zipObject(
-    Object.keys(metadataSetValues).map((k) =>
-      renamePkFields.includes(k) ? `${k}Pk` : k
-    ),
-    Object.values(metadataSetValues).map((v) => get(v, "value") || v)
-  );
-}
