+++ apps/admin-ui/src/component/reservations/EditTimeModal.tsx
@@ -17,12 +17,12 @@ import { differenceInMinutes, format } from "date-fns";
 import { ErrorBoundary } from "react-error-boundary";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { formatDuration, toUIDate } from "common/src/common/util";
-import { useNotification } from "app/context/NotificationContext";
-import { useModal } from "app/context/ModalContext";
-import { TimeChangeFormSchemaRefined, TimeFormSchema } from "app/schemas";
-import ControlledTimeInput from "../my-units/components/ControlledTimeInput";
-import ControlledDateInput from "../my-units/components/ControlledDateInput";
-import { BufferToggles } from "../my-units/BufferToggles";
+import { useNotification } from "@/context/NotificationContext";
+import { useModal } from "@/context/ModalContext";
+import { TimeChangeFormSchemaRefined, TimeFormSchema } from "@/schemas";
+import ControlledTimeInput from "@/component/ControlledTimeInput";
+import ControlledDateInput from "@/component/ControlledDateInput";
+import { BufferToggles } from "@/component/BufferToggles";
 import { useCheckCollisions } from "./requested/hooks";
 import { getNormalizedInterval, parseDateTimeSafe } from "@/helpers";
 import { formatDateTimeRange } from "@/common/util";
