+++ apps/admin-ui/src/hooks/index.ts
@@ -1,3 +1,6 @@
 export { useReservationUnitTypes } from "./useReservationUnitTypes";
 export { useReservationUnitOptions } from "./useReservationUnitOptions";
 export { useUnitOptions } from "./useUnitOptions";
+export { useOptions } from "./useOptions";
+export { useApplicationFields } from "./useApplicationFields";
+export { useGeneralFields } from "./useGeneralFields";
