+++ apps/admin-ui/src/hooks/useOptions.tsx
@@ -1,0 +1,31 @@
+import { sortBy } from "lodash";
+import { useOptionsQuery } from "@gql/gql-types";
+
+export function useOptions() {
+  const { data: optionsData } = useOptionsQuery();
+
+  const purpose = sortBy(
+    optionsData?.reservationPurposes?.edges || [],
+    "node.nameFi"
+  ).map((purposeType) => ({
+    label: purposeType?.node?.nameFi ?? "",
+    value: Number(purposeType?.node?.pk),
+  }));
+
+  const ageGroup = sortBy(
+    optionsData?.ageGroups?.edges || [],
+    "node.minimum"
+  ).map((group) => ({
+    label: `${group?.node?.minimum}-${group?.node?.maximum || ""}`,
+    value: Number(group?.node?.pk),
+  }));
+
+  const homeCity = sortBy(optionsData?.cities?.edges || [], "node.nameFi").map(
+    (cityType) => ({
+      label: cityType?.node?.nameFi ?? "",
+      value: Number(cityType?.node?.pk),
+    })
+  );
+
+  return { ageGroup, purpose, homeCity };
+}
