+++ apps/admin-ui/src/spa/my-units/MyUnits.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import { H1 } from "common/src/common/typography";
 import { Container } from "@/styles/layout";
-import { Filters } from "../Unit/Filters";
+import { Filters } from "@/component/Unit/Filters";
 import { HR } from "@/component/Table";
-import { UnitsDataLoader } from "../Unit/UnitsDataLoader";
+import { UnitsDataLoader } from "@/component/Unit/UnitsDataLoader";
 
 // NOTE copy pasta from Unit/Units.tsx
-function MyUnits() {
+export function MyUnits() {
   const { t } = useTranslation();
 
   return (
@@ -22,5 +22,3 @@ function MyUnits() {
     </Container>
   );
 }
-
-export default MyUnits;
