+++ apps/admin-ui/src/spa/my-units/ReservationPopupContent.tsx
@@ -4,12 +4,11 @@ import styled from "styled-components";
 import type { ReservationUnitReservationsFragment } from "@gql/gql-types";
 import { Permission } from "@/modules/permissionHelper";
 import { reservationUrl } from "@/common/urls";
-import { formatTime } from "@/common/util";
+import { formatTime, getReserveeName } from "@/common/util";
 import { truncate } from "@/helpers";
 import { DenseVerticalFlex } from "@/styles/layout";
-import { getReserveeName } from "../reservations/requested/util";
 import { CELL_BORDER } from "./const";
-import VisibleIfPermission from "../reservations/requested/VisibleIfPermission";
+import VisibleIfPermission from "@/component/reservations/requested/VisibleIfPermission";
 import { useTranslation } from "next-i18next";
 
 const MAX_POPOVER_COMMENT_LENGTH = 140;
