+++ apps/admin-ui/src/spa/my-units/UnitReservations.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { breakpoints } from "common/src/common/style";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
-import Loader from "../Loader";
-import Legend from "../reservations/requested/Legend";
+import Loader from "@/component/Loader";
+import Legend from "@/component/reservations/requested/Legend";
 import { legend } from "./eventStyleGetter";
 import { UnitCalendar } from "./UnitCalendar";
 import { useUnitResources } from "./hooks";
