+++ apps/admin-ui/src/spa/my-units/hooks/index.ts
@@ -1,80 +1,11 @@
-import { useMemo } from "react";
-import { sortBy } from "lodash";
-import { getReservationApplicationFields } from "common/src/reservation-form/util";
 import {
-  CustomerTypeChoice,
   ReservationTypeChoice,
   useReservationUnitsByUnitQuery,
-  useOptionsQuery,
-  type MetadataSetsFragment,
 } from "@gql/gql-types";
 import { toApiDate } from "common/src/common/util";
 import { base64encode, filterNonNullable } from "common/src/helpers";
 import { useNotification } from "@/context/NotificationContext";
 import { RELATED_RESERVATION_STATES } from "common/src/const";
-import { containsField } from "common/src/metaFieldsHelpers";
-
-export function useApplicationFields(
-  reservationUnit: MetadataSetsFragment,
-  reserveeType?: CustomerTypeChoice
-) {
-  return useMemo(() => {
-    const fields = filterNonNullable(
-      reservationUnit.metadataSet?.supportedFields
-    );
-
-    const type =
-      reserveeType != null && containsField(fields, "reserveeType")
-        ? reserveeType
-        : CustomerTypeChoice.Individual;
-
-    return getReservationApplicationFields({
-      supportedFields: fields,
-      reserveeType: type,
-    });
-  }, [reservationUnit.metadataSet?.supportedFields, reserveeType]);
-}
-
-export function useGeneralFields(reservationUnit: MetadataSetsFragment) {
-  return useMemo(() => {
-    const fields = filterNonNullable(
-      reservationUnit.metadataSet?.supportedFields
-    );
-    return getReservationApplicationFields({
-      supportedFields: fields,
-      reserveeType: "common",
-    }).filter((n) => n !== "reserveeType");
-  }, [reservationUnit.metadataSet?.supportedFields]);
-}
-
-export const useOptions = () => {
-  const { data: optionsData } = useOptionsQuery();
-
-  const purpose = sortBy(
-    optionsData?.reservationPurposes?.edges || [],
-    "node.nameFi"
-  ).map((purposeType) => ({
-    label: purposeType?.node?.nameFi ?? "",
-    value: Number(purposeType?.node?.pk),
-  }));
-
-  const ageGroup = sortBy(
-    optionsData?.ageGroups?.edges || [],
-    "node.minimum"
-  ).map((group) => ({
-    label: `${group?.node?.minimum}-${group?.node?.maximum || ""}`,
-    value: Number(group?.node?.pk),
-  }));
-
-  const homeCity = sortBy(optionsData?.cities?.edges || [], "node.nameFi").map(
-    (cityType) => ({
-      label: cityType?.node?.nameFi ?? "",
-      value: Number(cityType?.node?.pk),
-    })
-  );
-
-  return { ageGroup, purpose, homeCity };
-};
 
 export function useUnitResources(
   begin: Date,
