+++ apps/admin-ui/src/spa/my-units/recurring/commonStyling.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const ActionsWrapper = styled.div`
+  display: flex;
+  grid-column: 1 / -1;
+  gap: var(--spacing-m);
+  margin-top: var(--spacing-l);
+  margin-bottom: var(--spacing-l);
+  justify-content: end;
+`;
